Moved GameManager to root package and assigned name to the game thread
