+++ src/main/java/com/fabbroniko/environment/AudioManager.java
@@ -2,7 +2,7 @@ package com.fabbroniko.environment;
 
 import javax.sound.sampled.Clip;
 
-import com.fabbroniko.gamestatemanager.GameManager;
+import com.fabbroniko.GameManager;
 import com.fabbroniko.resource.ResourceManager;
 
 import java.util.Optional;
