+++ src/main/java/com/fabbroniko/scene/AbstractScene.java
@@ -2,7 +2,7 @@ package com.fabbroniko.scene;
 
 import com.fabbroniko.environment.AudioManager;
 import com.fabbroniko.environment.Dimension;
-import com.fabbroniko.gamestatemanager.GameManager;
+import com.fabbroniko.GameManager;
 import com.fabbroniko.main.Drawable;
 import com.fabbroniko.resource.ResourceManager;
