+++ src/main/java/com/fabbroniko/scene/WinScene.java
@@ -6,7 +6,7 @@ import java.awt.Color;
 
 import com.fabbroniko.environment.Dimension;
 import com.fabbroniko.environment.Position;
-import com.fabbroniko.gamestatemanager.GameManager;
+import com.fabbroniko.GameManager;
 
 /**
  * The WinScene is a very simple scene, it shows a "Level Completed" string on a simple background and it plays
